+++ docs/changes.rst
@@ -6,6 +6,14 @@ This is the changes log for the `badger_bits module
 <https://github.com/gradha/badger_bits>`_.  See the `README <../README.rst>`_.
 
 
+v0.2.4, 2015-01-17
+------------------
+
+* `Uses copy_dir_with_permissions() in bb_nake
+  <https://github.com/gradha/badger_bits/issues/10>`_.
+* `Added ini parameter to switch_to_gh_pages()
+  <https://github.com/gradha/badger_bits/issues/11>`_.
+
 v0.2.2, 2015-01-03
 ------------------